check_correct_function_call: refactor
Collect arguments in a struct.

Arguments to the function are now: the struct, the parsed arguments and CompileInfo.

Make a const map of predefined function specifications.

Call check_correct_function_call() at the beginning of the switch
statement while accessing the map with the function type; thus
remove the need to call in every case.