Use std::string_view
The string-view concept describes a data structure pointing to a str and
holding a length, which is not allowed to modify the string. This is
useful for passing around strings which would otherwise have been by
const-ref, and cutting them up without copying.

Some functions had to be reimplemented and a lot of things refactored
semi-automatically, look out for errors.