ast::parse_arit_expression: fix bugs
Previously, when pushing the number to ignore a bracket in an array like
s1 or s2, the current index in another array was nonsensesically pushed.
That was an oversight that I only discovered now after more bug testing.
Additionally, it was assumed that an expression with precedence could
take any preceding bracket expression. This was false and corrected.